+++ main.go
@@ -1,11 +1,9 @@
 package main
 
 import (
-	"encoding/json"
 	"errors"
 	"fmt"
 	"testcni/cni"
-	"testcni/consts"
 
 	_ "testcni/plugins/hostgw"
 	_ "testcni/plugins/vxlan/vxlan"
@@ -15,33 +13,6 @@ import (
 	"github.com/containernetworking/cni/pkg/version"
 	bv "github.com/containernetworking/plugins/pkg/utils/buildversion"
 )
-
-func getConfigs(args *skel.CmdArgs) *cni.PluginConf {
-	pluginConfig := &cni.PluginConf{}
-	if err := json.Unmarshal(args.StdinData, pluginConfig); err != nil {
-		utils.WriteLog("args.StdinData 转 pluginConfig 失败")
-		return nil
-	}
-	utils.WriteLog("这里的结果是: pluginConfig.Bridge", pluginConfig.Bridge)
-	utils.WriteLog("这里的结果是: pluginConfig.CNIVersion", pluginConfig.CNIVersion)
-	utils.WriteLog("这里的结果是: pluginConfig.Name", pluginConfig.Name)
-	utils.WriteLog("这里的结果是: pluginConfig.Subnet", pluginConfig.Subnet)
-	utils.WriteLog("这里的结果是: pluginConfig.Type", pluginConfig.Type)
-	utils.WriteLog("这里的结果是: pluginConfig.Mode", pluginConfig.Mode)
-	return pluginConfig
-}
-
-func getBaseInfo(plugin *cni.PluginConf) (mode string, cniVersion string) {
-	mode = plugin.Mode
-	if mode == "" {
-		mode = consts.MODE_HOST_GW
-	}
-	cniVersion = plugin.CNIVersion
-	if cniVersion == "" {
-		cniVersion = "0.3.0"
-	}
-	return mode, cniVersion
-}
 
 func cmdAdd(args *skel.CmdArgs) error {
 	utils.WriteLog("进入到 cmdAdd")
